requirements.txt
langchain==0.3.24
langchain-google-genai==2.1.3
python-dotenv==1.1.0
